Adds negation to multi-valued cells.
When you place a value in one cell, all peer cells with multiple values that include the placed value have that value negated, so that if you enter that value there it will be removed from the set of possibilities.  This is shown with an underline under the negated numbers.